Break components, types, and utils out of TaskEndpoints file
