+++ src/components/shared/Endpoints/types.ts
@@ -1,0 +1,11 @@
+import { Endpoint } from 'stores/ShardDetail/types';
+
+export interface TaskEndpointProps {
+    reactorAddress: string | undefined;
+    taskName: string;
+}
+
+export interface EndpointLinkProps {
+    endpoint: Endpoint;
+    hostName: string | null;
+}
